+++ lib/core.dart
@@ -1,6 +1,7 @@
-export 'src/builder/_.index.dart';
-export 'src/callback/_.index.dart';
-export 'src/exception/_.index.dart';
-export 'src/lifecycle/_.index.dart';
-export 'src/object/_.index.dart';
-export 'src/serialize/_.index.dart';
+export 'builder.dart';
+export 'callback.dart';
+export 'exception.dart';
+export 'lifecycle.dart';
+export 'object.dart';
+export 'serialize.dart';
+export 'stream.dart';