+++ lib/exception.dart
@@ -1,0 +1,1 @@
+export 'src/exception/_.index.dart';