+++ lib/lifecycle.dart
@@ -1,0 +1,1 @@
+export 'src/lifecycle/_.index.dart';