+++ lib/object.dart
@@ -1,0 +1,1 @@
+export 'src/object/_.index.dart';