+++ lib/redux.dart
@@ -1,1 +1,0 @@
-export 'package:redux/redux.dart';