+++ lib/serialize.dart
@@ -1,0 +1,1 @@
+export 'src/serialize/_.index.dart';