+++ lib/src/object/_.index.dart
@@ -1,4 +1,3 @@
-export 'object_as_stream.extension.dart';
 export 'categorizable.dart';
 export 'cloneable.dart';
 export 'copyable.dart';