+++ lib/stream.dart
@@ -1,0 +1,1 @@
+export 'src/stream/_.index.dart';