+++ lib/zam_core.dart
@@ -1,4 +1,4 @@
 export 'core.dart';
 export 'equatable.dart';
 export 'meta.dart';
-export 'redux.dart';
+export 'rxdart.dart';